+++ x/lockup/keeper/grpc_query.go
@@ -177,6 +177,21 @@ func (q Querier) NextLockID(goCtx context.Context, req *types.NextLockIDRequest)
 	return &types.NextLockIDResponse{LockId: nextLockID}, nil
 }
 
+// SyntheticLockupsByLockupID returns synthetic lockups by native lockup id.
+// Deprecated: use SyntheticLockupByLockupID instead.
+func (q Querier) SyntheticLockupsByLockupID(goCtx context.Context, req *types.SyntheticLockupsByLockupIDRequest) (*types.SyntheticLockupsByLockupIDResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "empty request")
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	synthLock, err := q.Keeper.GetSyntheticLockupByUnderlyingLockId(ctx, req.LockId)
+	if err != nil {
+		return nil, err
+	}
+	return &types.SyntheticLockupsByLockupIDResponse{SyntheticLocks: []types.SyntheticLock{synthLock}}, nil
+}
+
 // SyntheticLockupByLockupID returns synthetic lockup by native lockup id.
 func (q Querier) SyntheticLockupByLockupID(goCtx context.Context, req *types.SyntheticLockupByLockupIDRequest) (*types.SyntheticLockupByLockupIDResponse, error) {
 	if req == nil {